+++ src/mjlab/tasks/tracking/rl/runner.py
@@ -4,6 +4,7 @@ from typing import cast
 import torch
 import wandb
 from rsl_rl.env.vec_env import VecEnv
+from rsl_rl.utils import WandbLogWriter
 from torch import nn
 
 from mjlab.rl import RslRlVecEnvWrapper
@@ -94,11 +95,8 @@ class MotionTrackingOnPolicyRunner(MjlabOnPolicyRunner):
     policy_dir, filename, onnx_path = self._get_export_paths(path)
     try:
       self.export_policy_to_onnx(str(policy_dir), filename)
-      run_name: str = (
-        wandb.run.name
-        if self.logger.logger_type in ("wandb", "WandbLogWriter") and wandb.run
-        else "local"
-      )  # type: ignore[assignment]
+      is_wandb = isinstance(self.logger.writer, WandbLogWriter)
+      run_name: str = wandb.run.name if is_wandb and wandb.run else "local"  # type: ignore[assignment]
       metadata = get_base_metadata(self.env.unwrapped, run_name)
       motion_term = cast(
         MotionCommand, self.env.unwrapped.command_manager.get_term("motion")
@@ -110,10 +108,7 @@ class MotionTrackingOnPolicyRunner(MjlabOnPolicyRunner):
         }
       )
       attach_metadata_to_onnx(str(onnx_path), metadata)
-      if (
-        self.logger.logger_type in ("wandb", "WandbLogWriter")
-        and self.cfg["upload_model"]
-      ):
+      if is_wandb and self.cfg["upload_model"]:
         wandb.save(str(onnx_path), base_path=str(policy_dir))
         if self.registry_name is not None:
           wandb.run.use_artifact(self.registry_name)  # type: ignore